[Query] Use a minimal index pattern interface for es query  (#102364)
* Move JSON utils to utils package

* Imports from tests

* delete

* split package

* docs

* test

* test

* imports

* minimal index pattern

* move some functions out and use miniaml ip in all es-kuery

* docs

* docs

* rename

# Conflicts:
#	docs/development/plugins/data/public/kibana-plugin-plugins-data-public.iindexpattern.md
#	src/plugins/data/public/public.api.md
