--- docs/development/plugins/data/public/kibana-plugin-plugins-data-public.iindexpattern.fields.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [IIndexPattern](./kibana-plugin-plugins-data-public.iindexpattern.md) &gt; [fields](./kibana-plugin-plugins-data-public.iindexpattern.fields.md)
-
-## IIndexPattern.fields property
-
-<b>Signature:</b>
-
-```typescript
-fields: IFieldType[];
-```
