--- docs/development/plugins/data/public/kibana-plugin-plugins-data-public.iindexpattern.id.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [IIndexPattern](./kibana-plugin-plugins-data-public.iindexpattern.md) &gt; [id](./kibana-plugin-plugins-data-public.iindexpattern.id.md)
-
-## IIndexPattern.id property
-
-<b>Signature:</b>
-
-```typescript
-id?: string;
-```
