+++ docs/development/plugins/data/public/kibana-plugin-plugins-data-public.iindexpattern.md
@@ -12,7 +12,7 @@
 <b>Signature:</b>
 
 ```typescript
-export interface IIndexPattern 
+export interface IIndexPattern extends IndexPatternBase 
 ```
 
 ## Properties
@@ -20,9 +20,7 @@ export interface IIndexPattern
 |  Property | Type | Description |
 |  --- | --- | --- |
 |  [fieldFormatMap](./kibana-plugin-plugins-data-public.iindexpattern.fieldformatmap.md) | <code>Record&lt;string, SerializedFieldFormat&lt;unknown&gt; &#124; undefined&gt;</code> |  |
-|  [fields](./kibana-plugin-plugins-data-public.iindexpattern.fields.md) | <code>IFieldType[]</code> |  |
 |  [getFormatterForField](./kibana-plugin-plugins-data-public.iindexpattern.getformatterforfield.md) | <code>(field: IndexPatternField &#124; IndexPatternField['spec'] &#124; IFieldType) =&gt; FieldFormat</code> | Look up a formatter for a given field |
-|  [id](./kibana-plugin-plugins-data-public.iindexpattern.id.md) | <code>string</code> |  |
 |  [intervalName](./kibana-plugin-plugins-data-public.iindexpattern.intervalname.md) | <code>string &#124; null</code> |  |
 |  [timeFieldName](./kibana-plugin-plugins-data-public.iindexpattern.timefieldname.md) | <code>string</code> |  |
 |  [title](./kibana-plugin-plugins-data-public.iindexpattern.title.md) | <code>string</code> |  |
