+++ src/plugins/data/common/es_query/es_query/build_es_query.ts
@@ -10,9 +10,9 @@ import { groupBy, has, isEqual } from 'lodash';
 import { buildQueryFromKuery } from './from_kuery';
 import { buildQueryFromFilters } from './from_filters';
 import { buildQueryFromLucene } from './from_lucene';
-import { IIndexPattern } from '../../index_patterns';
 import { Filter } from '../filters';
 import { Query } from '../../query/types';
+import { IndexPatternBase } from './types';
 
 export interface EsQueryConfig {
   allowLeadingWildcards: boolean;
@@ -36,7 +36,7 @@ function removeMatchAll<T>(filters: T[]) {
  * config contains dateformat:tz
  */
 export function buildEsQuery(
-  indexPattern: IIndexPattern | undefined,
+  indexPattern: IndexPatternBase | undefined,
   queries: Query | Query[],
   filters: Filter | Filter[],
   config: EsQueryConfig = {
