+++ src/plugins/data/common/es_query/es_query/filter_matches_index.ts
@@ -6,15 +6,16 @@
  * Side Public License, v 1.
  */
 
-import { IIndexPattern, IFieldType } from '../../index_patterns';
+import { IFieldType } from '../../index_patterns';
 import { Filter } from '../filters';
+import { IndexPatternBase } from './types';
 
 /*
  * TODO: We should base this on something better than `filter.meta.key`. We should probably modify
  * this to check if `filter.meta.index` matches `indexPattern.id` instead, but that's a breaking
  * change.
  */
-export function filterMatchesIndex(filter: Filter, indexPattern?: IIndexPattern | null) {
+export function filterMatchesIndex(filter: Filter, indexPattern?: IndexPatternBase | null) {
   if (!filter.meta?.key || !indexPattern) {
     return true;
   }
