+++ src/plugins/data/common/es_query/es_query/from_filters.ts
@@ -10,7 +10,7 @@ import { isUndefined } from 'lodash';
 import { migrateFilter } from './migrate_filter';
 import { filterMatchesIndex } from './filter_matches_index';
 import { Filter, cleanFilter, isFilterDisabled } from '../filters';
-import { IIndexPattern } from '../../index_patterns';
+import { IndexPatternBase } from './types';
 import { handleNestedFilter } from './handle_nested_filter';
 
 /**
@@ -45,7 +45,7 @@ const translateToQuery = (filter: Filter) => {
 
 export const buildQueryFromFilters = (
   filters: Filter[] = [],
-  indexPattern: IIndexPattern | undefined,
+  indexPattern: IndexPatternBase | undefined,
   ignoreFilterIfFieldNotInIndex: boolean = false
 ) => {
   filters = filters.filter((filter) => filter && !isFilterDisabled(filter));
