+++ src/plugins/data/common/es_query/filters/index.ts
@@ -14,10 +14,8 @@ export * from './custom_filter';
 export * from './exists_filter';
 export * from './geo_bounding_box_filter';
 export * from './geo_polygon_filter';
-export * from './get_display_value';
 export * from './get_filter_field';
 export * from './get_filter_params';
-export * from './get_index_pattern_from_filter';
 export * from './match_all_filter';
 export * from './meta_filter';
 export * from './missing_filter';
