+++ src/plugins/data/common/es_query/filters/phrases_filter.ts
@@ -9,7 +9,8 @@
 import { Filter, FilterMeta } from './meta_filter';
 import { getPhraseScript } from './phrase_filter';
 import { FILTERS } from './index';
-import { IIndexPattern, IFieldType } from '../../index_patterns';
+import { IFieldType } from '../../index_patterns';
+import { IndexPatternBase } from '../es_query';
 
 export type PhrasesFilterMeta = FilterMeta & {
   params: string[]; // The unformatted values
@@ -34,7 +35,7 @@ export const getPhrasesFilterField = (filter: PhrasesFilter) => {
 export const buildPhrasesFilter = (
   field: IFieldType,
   params: any[],
-  indexPattern: IIndexPattern
+  indexPattern: IndexPatternBase
 ) => {
   const index = indexPattern.id;
   const type = FILTERS.PHRASES;
