+++ src/plugins/data/common/es_query/kuery/ast/ast.ts
@@ -10,10 +10,10 @@ import { JsonObject } from '@kbn/common-utils';
 import { nodeTypes } from '../node_types/index';
 import { KQLSyntaxError } from '../kuery_syntax_error';
 import { KueryNode, DslQuery, KueryParseOptions } from '../types';
-import { IIndexPattern } from '../../../index_patterns/types';
 
 // @ts-ignore
 import { parse as parseKuery } from './_generated_/kuery';
+import { IndexPatternBase } from '../..';
 
 const fromExpression = (
   expression: string | DslQuery,
@@ -65,7 +65,7 @@ export const fromKueryExpression = (
  */
 export const toElasticsearchQuery = (
   node: KueryNode,
-  indexPattern?: IIndexPattern,
+  indexPattern?: IndexPatternBase,
   config?: Record<string, any>,
   context?: Record<string, any>
 ): JsonObject => {
