+++ src/plugins/data/common/es_query/kuery/functions/or.ts
@@ -7,7 +7,7 @@
  */
 
 import * as ast from '../ast';
-import { IIndexPattern, KueryNode } from '../../..';
+import { IndexPatternBase, KueryNode } from '../../..';
 
 export function buildNodeParams(children: KueryNode[]) {
   return {
@@ -17,7 +17,7 @@ export function buildNodeParams(children: KueryNode[]) {
 
 export function toElasticsearchQuery(
   node: KueryNode,
-  indexPattern?: IIndexPattern,
+  indexPattern?: IndexPatternBase,
   config: Record<string, any> = {},
   context: Record<string, any> = {}
 ) {
