+++ src/plugins/data/common/es_query/kuery/functions/utils/get_full_field_name_node.ts
@@ -7,11 +7,11 @@
  */
 
 import { getFields } from './get_fields';
-import { IIndexPattern, IFieldType, KueryNode } from '../../../..';
+import { IndexPatternBase, IFieldType, KueryNode } from '../../../..';
 
 export function getFullFieldNameNode(
   rootNameNode: any,
-  indexPattern?: IIndexPattern,
+  indexPattern?: IndexPatternBase,
   nestedPath?: string
 ): KueryNode {
   const fullFieldNameNode = {
