+++ src/plugins/data/common/es_query/kuery/node_types/function.ts
@@ -9,7 +9,7 @@
 import _ from 'lodash';
 
 import { functions } from '../functions';
-import { IIndexPattern, KueryNode } from '../../..';
+import { IndexPatternBase, KueryNode } from '../../..';
 import { FunctionName, FunctionTypeBuildNode } from './types';
 
 export function buildNode(functionName: FunctionName, ...args: any[]) {
@@ -45,7 +45,7 @@ export function buildNodeWithArgumentNodes(
 
 export function toElasticsearchQuery(
   node: KueryNode,
-  indexPattern?: IIndexPattern,
+  indexPattern?: IndexPatternBase,
   config?: Record<string, any>,
   context?: Record<string, any>
 ) {
