+++ src/plugins/data/common/es_query/kuery/node_types/types.ts
@@ -11,8 +11,8 @@
  */
 
 import { JsonValue } from '@kbn/common-utils';
-import { IIndexPattern } from '../../../index_patterns';
 import { KueryNode } from '..';
+import { IndexPatternBase } from '../..';
 
 export type FunctionName =
   | 'is'
@@ -30,7 +30,7 @@ interface FunctionType {
   buildNodeWithArgumentNodes: (functionName: FunctionName, args: any[]) => FunctionTypeBuildNode;
   toElasticsearchQuery: (
     node: any,
-    indexPattern?: IIndexPattern,
+    indexPattern?: IndexPatternBase,
     config?: Record<string, any>,
     context?: Record<string, any>
   ) => JsonValue;
