+++ src/plugins/data/public/query/filter_manager/index.ts
@@ -11,3 +11,5 @@ export { FilterManager } from './filter_manager';
 export { mapAndFlattenFilters } from './lib/map_and_flatten_filters';
 export { onlyDisabledFiltersChanged } from './lib/only_disabled';
 export { generateFilters } from './lib/generate_filters';
+export { getDisplayValueFromFilter } from './lib/get_display_value';
+export { getIndexPatternFromFilter } from './lib/get_index_pattern_from_filter';
