+++ src/plugins/data/public/query/filter_manager/lib/get_display_value.ts
@@ -7,9 +7,8 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { IIndexPattern } from '../..';
+import { Filter, IIndexPattern } from '../../../../common';
 import { getIndexPatternFromFilter } from './get_index_pattern_from_filter';
-import { Filter } from '../filters';
 
 function getValueFormatter(indexPattern?: IIndexPattern, key?: string) {
   // checking getFormatterForField exists because there is at least once case where an index pattern
