+++ src/plugins/data/public/query/filter_manager/lib/get_index_pattern_from_filter.test.ts
